Move tests to main files
